Make clear screen binding work again in Tmux
This was broken since moving to use the TPM version of the
`vim-tmux-navigator` plugin (rather than copying the lines of this
plugin into my Tmux config), due to `is_vim` no longer being defined
within my config. The solution is to inline the definition of `is_vim`
again, so that the correct key is sent when in Tmux as well as when in
Vim.