Tweak ctags config; get TSX tags working (fixes #4)
Switch to ctags directory rather than config file, which is apparently
the correct location for ctags configs now.

Can then delete Elm config, as this seems to be built-in to latest
version of ctags (was getting an error `ctags: Language "Elm" already
defined` with this still present at least).

And then adding this line gets ctags working for `.tsx` files (see
https://github.com/universal-ctags/ctags/issues/2127#issuecomment-507942820).