Use path rather than env var in Terminator config
Using an environment variable seemed to be causing this config file to
fail to load when a terminal is launched after logging in again, due to
this not being defined yet, which then meant only default Terminator
settings were being used in all terminals. Switching to just use a path
seems to resolve this.