Include current session name in secondary Tmux session
This makes it clearer which other session each secondary Tmux session is
associated with, and also means any special setup behaviour that should
occur for sessions with particular names (such as sourcing a virtualenv
for `rescale` sessions) can also easily be triggered for the secondary
session, by just matching on this part of the name.

As part of this commit also bump wait for secondary session to launch in
`RunTests`, to allow more time for any additional behaviour to run.