Correct `CrRunCurrentFile` to always run original file
Rather than have this execute whatever file the cursor is in when the
mapping is run. Usually these have always been the same when I've used
this before, so never noticed this issue, but this is not always the
case; e.g. when working on a more complex script that imports and uses
another file, always want <CR> to execute the main script.