Split shell functions/aliases into more files (resolves #23)
To stop `zsh/lib/main.sh` being a junk drawer for all sorts of things.
Where a function/alias is related to a particular niche area, move it to
a file specific to that area; a few fairly general purpose utilities
remain in a renamed `zsh/lib/utils.sh` file.