Have `rake` and `rails` run at Git root using shims
Have these always run at the Git root, using the shims within `bin/` (in
Rake's case only if this is present, to account for non-Rails projects).
This is useful so these always run in the same way no matter which
directory within a repo they are run for, and always account for
additional behaviour that might be added to these shims.