Symlink new TSX filetype snippets to old
For some reason (probably with changing some plugin) the name of the TSX
filetype has changed recently. Therefore link the new filetype snippets
file to the old name, so the snippets are picked up for both filetypes
(might as well do this rather than just renaming, in case ever go back
to the old name for any reason).