Split `gl` and `git_log` out to separate scripts
So that `gl` can be used in places where aliases/functions are not
available (like other scripts or Git aliases); `git_log` in turn also
needs to be a script so it can be used by `gl`. The other `git log`
aliases can stay where they are as don't need them elsewhere.