Fix `git sed` to work when non-text files present
Previously this would error like this when certain non-text files are
present in the repo (and have not already been filtered out by the
`file_filter`):

```bash
sed: couldn't edit some_file: not a regular file
```

In the case I encountered, the "file" in question was actually a
submodule in a bad state, but possibly this can occur in other
situations too.

This commit fixes this so we're only searching text files, though the
fix is still a bit hacky and any bad file with "text" in the name would
still cause issues.