Make generic standard project Tmuxinator config
Starts standard windows and panes, with any project-specific setup in
central pane, and session and windows appropriately named. This can then
replace `npm-project` config, `calm` config, and serve as a generic
config for simple projects with similar or no setup (and easy to slot in
new setup types).

Also add alias (as a shell function) to make using this easier.