Disable `vim-closetag` in markdown files
Having this enabled in markdown files breaks repeating inserting a `>`
character (such as when inserting quotes), which is very annoying, and I
very rarely Write HTML within markdown anyway.

Also not sure if the  `*.js,*.jsx,*.tsx` part here is actually required
since it is included in `g:closetag_xhtml_filenames` below and is not
included here in the docs
(https://github.com/alvan/vim-closetag?tab=readme-ov-file#options) - but
will just leave this as before for now.