Use `less -F` option to simplify `tree` aliases
This causes `less` to just print to screen if the output will fit on one
screen, and page as normal if it will not. This means `treel` is
redundant as my `tree` alias can now do the same thing, and this alias
can also now depend on `treeall` directly without indirection.