Add snippet for adding more useful TODO comments
I used to think including the author and date in TODO comments was
unnecessary, as `git blame` is sufficient to find these out. But code
moves over time, and so trying to find the true origin of something can
sometimes require a fair amount of archaeology. This way this info is
always up-front, and you can tell exactly how stale or who to ask about
a particular TODO.